+++ activerecord/CHANGELOG.md
@@ -1,3 +1,9 @@
+*   Add `:enum_prefix`/`:enum_suffix` option to `enum` definition.
+
+    Fixes #17511 and #17415
+
+    *Igor Kapkov*
+
 *   Correctly handle decimal arrays with defaults in the schema dumper.
 
     Fixes #20515.